+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Change posts more text customizer control label to 'More link text'
 - Rename `languages/` directory to `lang/`
 - Move reusable methods in `Setups\MetaBoxes\Layout` to `Setups\MetaBoxes\AbstractMetaBox`
+- Prefix registered thumbnail sizes with `jentil-`
 
 ## 0.8.0 - 2018-08-24